+++ tests/scalers/azure/azure_queue_aad_pod_identity/azure_queue_aad_pod_identity_test.go
@@ -162,7 +162,9 @@ func getTemplateData() (templateData, []Template) {
 			SecretName:       secretName,
 			Connection:       base64ConnectionString,
 			DeploymentName:   deploymentName,
+			TriggerAuthName:  triggerAuthName,
 			ScaledObjectName: scaledObjectName,
+			AccountName:      accountName,
 			QueueName:        queueName,
 		}, []Template{
 			{Name: "secretTemplate", Config: secretTemplate},